Drop leading \ from imported classes.
The PHP docs for use discourage use of a leading \ for
the class in a use statement. However it is still valid
for it to be specified  and is equivalent to the case
where it is not present.

Previous Doctrine behaviour was to treat them as
different, but that led to confusing behaviour:
e.g. you needed to register an autoloader both with
and without the leading \, just in case developers
erroniously included a leading \ in their use statements.

This change strips the leading \ in parseUseStatement,
normalizing use statements to never have the leading \.